Create new plugin for monitoring inode usage
Created a s new sensu plugin for monitoring inode usage.  It is basiclly a port of the community plugin \'check-disk\' with the additional \'-i\' flag and some name changes.
It is functionally identical though.  I would at some point like to add some validation for the input variables as an un caught exception is thrown when a non-valid input is given.