Add monitor to check for the status of a raid array
This monitor will read '/proc/mdstat' and parse it for several values.

[n/m] where n is the total number of devices and m is the number of working devics.

[UUU_UU] where U means the specific component is up and _ means that it is down
It will also check for a recovery status in the case of a failed component

The alerting breaks down as follows

[green]:
  n == m
  no recovery state
[yellow]
  n > m
  recovery state
[red]
  n > m
  no recovery state
  array is !active

When defining active using mdstats, active implies that the device is up.  It can be in one of
several states such as sync, clean, etc