Fix portability issue and add docs
I added a documentation header to the monitor with links as to where further info can be found

I added a sanity check to confirm that '/proc/mdstat' was in fact present before trying to open it.
If it is not present then the script will exit gracefully saying the file is not present.
If the file is present but raid is not being used then the script will exit gracefully as the regex
  will return nothing